stash: Prevent ID reuse for compute instances (#13110)
The default behavior for ID selection in stash collections is to take
the maximum of all current IDs and add one. This means that after
deleting an item in a stash collection, it's ID can be reused, if the
collection is using the default behavior.

This is problematic for compute instances due to how TAIL is
implemented. TAIL creates a temporary sink that stores the ID of the
compute instance that it was created in. When the TAIL is canceled,
the associated sink is deleted from the compute instance, if the
compute instance still exists. The compute instance ID is used to
identify if the compute instance exists. The TAIL can outlive the
compute instance if the compute instance is deleted while the TAIL is
running. If the compute instance is deleted and new one is created,
then the new compute instance may share the same ID as the old one. The
TAIL will then mistakenly try and delete the sink from this new compute
instance and cause a panic.

This commit changes ID selection for compute instances so that they are
always monotonically increasing and no ID is ever reused. The
implementation is based on GlobalId allocation.

Fixes #13100

+++ src/coord/src/catalog/builtin_table_updates.rs
@@ -107,7 +107,8 @@ impl CatalogState {
         let id = self.compute_instances_by_name[name];
         BuiltinTableUpdate {
             id: self.resolve_builtin_table(&MZ_CLUSTERS),
-            row: Row::pack_slice(&[Datum::Int64(id), Datum::String(&name)]),
+            // TODO(jkosh44) when Uint64 is supported change below to Datum::Uint64
+            row: Row::pack_slice(&[Datum::Int64(id as i64), Datum::String(&name)]),
             diff,
         }
     }
@@ -134,7 +135,8 @@ impl CatalogState {
         BuiltinTableUpdate {
             id: self.resolve_builtin_table(&MZ_CLUSTER_REPLICAS_BASE),
             row: Row::pack_slice(&[
-                Datum::Int64(compute_instance_id),
+                // TODO(jkosh44) when Uint64 is supported change below to Datum::Uint64
+                Datum::Int64(compute_instance_id as i64),
                 Datum::Int64(id),
                 Datum::String(&name),
                 Datum::from(size),
@@ -384,7 +386,8 @@ impl CatalogState {
                     Datum::String(name),
                     Datum::String(connection.name()),
                     Datum::String(self.is_volatile(id).as_str()),
-                    Datum::Int64(sink.compute_instance),
+                    // TODO(jkosh44) when Uint64 is supported change below to Datum::Uint64
+                    Datum::Int64(sink.compute_instance as i64),
                 ]),
                 diff,
             });
@@ -418,7 +421,8 @@ impl CatalogState {
                 Datum::String(name),
                 Datum::String(&index.on.to_string()),
                 Datum::String(self.is_volatile(id).as_str()),
-                Datum::Int64(index.compute_instance),
+                // TODO(jkosh44) when Uint64 is supported change below to Datum::Uint64
+                Datum::Int64(index.compute_instance as i64),
             ]),
             diff,
         });
